Изменение порядкового номера у питомца

+++ backend/src/PetHelp.Application/VolunteerManagement/RemoveFile/RemoveFileHandler.cs
@@ -1,6 +1,8 @@
 ﻿using CSharpFunctionalExtensions;
+using FluentValidation;
 using Microsoft.Extensions.Logging;
-using PetHelp.Application.Providers;
+using PetHelp.Application.Extensions;
+using PetHelp.Application.FileProvider;
 using PetHelp.Domain.Shared;
 
 namespace PetHelp.Application.VolunteerManagement.RemoveFile;
@@ -9,19 +11,25 @@ public class RemoveFileHandler
 {
     private readonly IFileProvider _fileProvider;
     private readonly ILogger<RemoveFileHandler> _logger;
+    private readonly IValidator<RemoveFileCommand> _validator;
 
     public RemoveFileHandler(
         IFileProvider fileProvider,
-        ILogger<RemoveFileHandler> logger)
+        ILogger<RemoveFileHandler> logger, IValidator<RemoveFileCommand> validator)
     {
         _fileProvider = fileProvider;
         _logger = logger;
+        _validator = validator;
     }
 
     public async Task<Result<string, ErrorList>> Handle(
         RemoveFileCommand command,
         CancellationToken cancellationToken)
     {
+        var validationResult = await _validator.ValidateAsync(command, cancellationToken);
+        if (validationResult.IsValid == false)
+            return validationResult.ToErrorList();
+        
         var result = await _fileProvider.RemoveFile(
             command.BucketName,
             command.ObjectName,
@@ -31,6 +39,8 @@ public class RemoveFileHandler
         {
             return result.Error.ToErrorList();
         }
+        
+        _logger.LogInformation("Remove object with {ObjectName}", command.ObjectName);
 
         return result.Value;
     }
